x/rdex/types/message_rm_provider.go
package types

import (
	sdk "github.com/cosmos/cosmos-sdk/types"
	sdkerrors "github.com/cosmos/cosmos-sdk/types/errors"
)

const TypeMsgRmProvider = "rm_provider"

var _ sdk.Msg = &MsgRmProvider{}

func NewMsgRmProvider(creator string, userAddress string) *MsgRmProvider {
	return &MsgRmProvider{
		Creator:     creator,
		UserAddress: userAddress,
	}
}

func (msg *MsgRmProvider) Route() string {
	return RouterKey
}

func (msg *MsgRmProvider) Type() string {
	return TypeMsgRmProvider
}

func (msg *MsgRmProvider) GetSigners() []sdk.AccAddress {
	creator, err := sdk.AccAddressFromBech32(msg.Creator)
	if err != nil {
		panic(err)
	}
	return []sdk.AccAddress{creator}
}

func (msg *MsgRmProvider) GetSignBytes() []byte {
	bz := ModuleCdc.MustMarshalJSON(msg)
	return sdk.MustSortJSON(bz)
}

func (msg *MsgRmProvider) ValidateBasic() error {
	_, err := sdk.AccAddressFromBech32(msg.Creator)
	if err != nil {
		return sdkerrors.Wrapf(sdkerrors.ErrInvalidAddress, "invalid creator address (%s)", err)
	}
	_, err = sdk.AccAddressFromBech32(msg.UserAddress)
	if err != nil {
		return sdkerrors.Wrapf(sdkerrors.ErrInvalidAddress, "invalid provider address (%s)", err)
	}
	return nil
}
